+++ R/TTest.R
@@ -5,7 +5,7 @@
 #' @param statistic wrapped T-test
 #' @param ... not used, force use of named binding for later arguments
 #' @param format if set the format to return ("html", "latex", "markdown", "ascii")
-#' @param statDigits integer number of digits to show in summaries (not yet implemented).
+#' @param statDigits integer number of digits to show in summaries.
 #' @param sigDigits integer number of digits to show in significances.
 #' @param pLargeCutoff value to declare non-significance at or above.
 #' @param pSmallCutoff smallest value to print
@@ -31,6 +31,7 @@ render.sigr_ttest <- function(statistic,
   }
   fsyms <- syms[format,]
   tt <- statistic$tt
+  stat_format_str <- paste0('%.',statDigits,'g')
   pString <- render(wrapSignificance(tt$p.value,
                                      symbol='p'),
                     format=format,
@@ -39,9 +40,9 @@ render.sigr_ttest <- function(statistic,
   formatStr <- paste0(fsyms['startB'],tt$method,fsyms['endB'],
                      ', ',tt$alternative,
                      ': (',fsyms['startI'],'t',fsyms['endI'],
-                     '=',paste(sprintf('%.2g',tt$statistic),collapse=','),
+                     '=',paste(sprintf(stat_format_str,tt$statistic),collapse=','),
                      ', ',fsyms['startI'],'df',fsyms['endI'],'=',
-                     sprintf('%.2g',tt$parameter),', ',
+                     sprintf(stat_format_str,tt$parameter),', ',
                      pString,').')
   formatStr
 }
@@ -92,6 +93,12 @@ wrapTTest.htest <- function(x,
 #' @param Column1Name character column 1 name
 #' @param Column2Name character column 2 name
 #' @param ... extra arguments passed to ttest
+#' @param y passed to \code{\link[stats]{t.test}}
+#' @param alternative passed to \code{\link[stats]{t.test}}
+#' @param mu passed to \code{\link[stats]{t.test}}
+#' @param paired passed to \code{\link[stats]{t.test}}
+#' @param var.equal passed to \code{\link[stats]{t.test}}
+#' @param conf.level passed to \code{\link[stats]{t.test}}
 #' @param na.rm logical, if TRUE remove NA values
 #' @return formatted string and fields
 #'
@@ -110,6 +117,10 @@ wrapTTest.data.frame <- function(x,
                                  Column1Name,
                                  Column2Name,
                                  ...,
+                                 y = NULL,
+                                 alternative = c("two.sided", "less", "greater"),
+                                 mu = 0, paired = FALSE, var.equal = FALSE,
+                                 conf.level = 0.95,
                                  na.rm= FALSE) {
   if(!'data.frame' %in% class(x)) {
     stop('sigr::wrapTTest expected class data.frame')
@@ -129,11 +140,77 @@ wrapTTest.data.frame <- function(x,
     c2 <- c2[goodPosns]
   }
   n <- length(c1)
-  tt <- t.test(c1,c2,...)
+  tt <- stats::t.test(c1,c2,
+                      alternative = c("two.sided", "less", "greater"),
+                      mu = 0, paired = FALSE, var.equal = FALSE,
+                      conf.level = 0.95, ...)
   r <- list(tt=tt,
             test='t.test',
             Column1Name=Column1Name,
             Column2Name=Column2Name,
+            n=n,
+            nNA=nNA)
+  class(r) <- c('sigr_ttest', 'sigr_statistic')
+  r
+}
+
+
+#' Wrap t.test (difference in means by group).
+#'
+#' @param x numeric population 1
+#' @param pop2 numeric population 2
+#' @param ... extra arguments passed to ttest
+#' @param y passed to \code{\link[stats]{t.test}}
+#' @param alternative passed to \code{\link[stats]{t.test}}
+#' @param mu passed to \code{\link[stats]{t.test}}
+#' @param paired passed to \code{\link[stats]{t.test}}
+#' @param var.equal passed to \code{\link[stats]{t.test}}
+#' @param conf.level passed to \code{\link[stats]{t.test}}
+#' @param na.rm logical, if TRUE remove NA values
+#' @return formatted string and fields
+#'
+#' @examples
+#'
+#' d <- data.frame(x=c(1,2,3,4,5,6,7,7),
+#'                 y=c(1,1,2,2,3,3,4,4))
+#' render(wrapTTest(d$x, d$y), pLargeCutoff=1)
+#' # confirm p not order depedent
+#' render(wrapTTest(d$y, d$x),pLargeCutoff=1)
+#'
+#' @importFrom stats t.test
+#'
+#' @export
+wrapTTest.numeric <- function(x,
+                              pop2,
+                              ...,
+                              y = NULL,
+                              alternative = c("two.sided", "less", "greater"),
+                              mu = 0, paired = FALSE, var.equal = FALSE,
+                              conf.level = 0.95,
+                              na.rm= FALSE) {
+  if(!is.numeric(x)) {
+    stop("sigr::wrapTTest expected x to be numeric")
+  }
+  if(!is.numeric(pop2)) {
+    stop("sigr::wrapTTest expected pop2 to be numeric")
+  }
+  c1 <- x
+  c2 <- pop2
+  nNA <- sum(is.na(c1) | is.na(c2))
+  if(na.rm) {
+    goodPosns <- (!is.na(c1)) & (!is.na(c2))
+    c1 <- c1[goodPosns]
+    c2 <- c2[goodPosns]
+  }
+  n <- length(c1)
+  tt <- stats::t.test(c1,c2,
+                      alternative = c("two.sided", "less", "greater"),
+                      mu = 0, paired = FALSE, var.equal = FALSE,
+                      conf.level = 0.95, ...)
+  r <- list(tt=tt,
+            test='t.test',
+            Column1Name='x',
+            Column2Name='pop2',
             n=n,
             nNA=nNA)
   class(r) <- c('sigr_ttest', 'sigr_statistic')